Move metric/byte prefixes to better place in file structure
